Add LICENSE and Hex metadata, set vsn to 1.1.0
Note that the original code did not have a LICENSE and I (@yurrriq) am
adding one now. I'll need to confirm with @knutin that this is
acceptable. I've added The MIT License, which is the same as in
knutin/elli, including

    Copyright (c) 2012 Knut Nesheim

in recognition of Knut's original work.

Hopefully this is alright. I'm going with an "ask for forgiveness rather
than permission approach" here...